test: update docker image
